% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdata.R
\name{asr}
\alias{asr}
\title{Popdata ASR figures}
\usage{
asr(
  indicator = c("refugees", "refugeeLike", "returnees", "demographics", "idp", "rsd",
    "stateless", "vda", "other", "hostcommunity"),
  year = 2020
)
}
\arguments{
\item{indicator}{character, the indicator}

\item{year}{integer, the year}
}
\value{
A tibble
}
\description{
Popdata ASR figures
}
