% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdata.R
\name{mysr}
\alias{mysr}
\title{Popdata MYSR figures}
\usage{
mysr(
  indicator = c("refugees", "returnees", "idp", "rsd", "stateless"),
  year = 2020
)
}
\arguments{
\item{indicator}{character, the indicator}

\item{year}{integer, the year}
}
\value{
A tibble
}
\description{
Popdata MYSR figures
}
