% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdata.R
\name{pd_augment}
\alias{pd_augment}
\title{Augment data-frame with country metadata}
\usage{
pd_augment(data, col, prefix = NULL)
}
\arguments{
\item{data}{tibble, dataset to augment}

\item{col}{unquoted expression, column containing UNHCR code}

\item{prefix}{character, string to prepend to metadata column names}
}
\value{
a tibble
}
\description{
Augment data-frame with country metadata
}
