% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pd_cache}
\alias{pd_cache}
\alias{pd_cache_set_dir}
\alias{pd_cache_get_dir}
\alias{pd_cache_list}
\alias{pd_cache_delete}
\alias{pd_cache_clear}
\title{Caching POPDATA downloaded files}
\usage{
pd_cache_set_dir(path)

pd_cache_get_dir()

pd_cache_list()

pd_cache_delete(file)

pd_cache_clear()
}
\arguments{
\item{path}{character, directory to set}

\item{file}{Character, the file to delete}
}
\value{
the cache directory

the cache directory

list of files in the cache
}
\description{
Manage cached POPDATA downloaded files
}
\details{
The default cache directory is
\verb{~/.cache/R/popdata}, but you can set
your own path using \code{pd_cache_set_dir()}
}
\examples{
\dontrun{
pd_cache
## change the default cache directory
tmp <- tempdir()
pd_cache_set_dir(tmp)

## print current cache directory
pd_cache_get_dir()

## List available files in the current cache directory
pd_cache_list()

l <- pd_cache_list()[1] ## get the first file
pd_cache_delete(l) ## delete it

pd_cache_clear() ## delete all cached files
}
}
