% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd_countries}
\alias{pd_countries}
\title{Country metadata}
\format{
A data-frame with 226 rows and 15 variables:
\describe{
\item{bureau}{The UNHCR bureau name.}
\item{code}{The three character UNHCR code for each country (e.g. "ICO"). Note that to support the data, two non-standard codes are used for which no ISO codes are available: a. STA for Stateless; b. UKN for Unknown/other.}
\item{iso}{The three character ISO code for each country (e.g. "CIV").}
\item{iso2}{The two character ISO code for each country (e.g. "CI").}
\item{name}{The country name.}
\item{nameOrigin}{The name of the country when presented as a country of origin.  Typically this is the same as the country name.}
\item{nameLong}{The full country name.}
\item{nameShort}{The official abbreviated country name.}
\item{nameFormal}{The official country name.}
\item{nationality}{The nationality of individuals from the country.}
\item{majorArea}{The UNSD region.}
\item{region}{The UNSD sub-region.}
\item{nameFr}{The official country name in French.}
\item{majorAreaFr}{The UNSD region in French.}
\item{regionFr}{The UNSD sub-region in French.}
}
}
\source{
\url{https://api.unhcr.org/population/v1/}
}
\usage{
pd_countries
}
\description{
Information on the countries covered by the UNHCR Population Statistics Reference platform.
}
\keyword{datasets}
