% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdata.R
\name{planning}
\alias{planning}
\title{Popdata Planning figures}
\usage{
planning(indicator = c("coo", "ppg", "specific"), year = 2020)
}
\arguments{
\item{indicator}{character, the indicator}

\item{year}{integer, the year}
}
\value{
A tibble
}
\description{
Popdata Planning figures
}
