% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdata.R
\name{popdata}
\alias{popdata}
\title{popdata() function by Fang Jiang
This function takes input from the user on allele names and population size
It then prints out the genetic variation data}
\usage{
popdata(allele1, allele2, popsize)
}
\description{
popdata() function by Fang Jiang
This function takes input from the user on allele names and population size
It then prints out the genetic variation data
}
