% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{popdata_setup}
\alias{popdata_setup}
\title{Popdata credential}
\usage{
popdata_setup(username = NULL, password = NULL, cookie = NULL)
}
\arguments{
\item{username}{character, popdata user name}

\item{password}{character, popdata password}

\item{cookie}{character, popdata cookie}
}
\description{
Popdata credential
}
