% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight_samp.R
\name{reweight_data}
\alias{reweight_data}
\title{This function allows for the reweighting of any dataset
with information on age (in 10 year increments from 15 to 65),
race (black, hispanic, or other),
and region of Washington to match the ACS based joint age / race
distribution and proportion of MSM in each region.}
\usage{
reweight_data(
  obs_data,
  age_c_name = "age.grp",
  race_c_name = "race",
  reg_c_name = "region",
  ego_id_name = "ego.id"
)
}
\arguments{
\item{obs_data}{The observed data for which weights are wanted.}

\item{age_c_name}{The name of the age column in the data frame.}

\item{race_c_name}{The name of the race column in the data frame.}

\item{reg_c_name}{The name of the region column in the data frame.}

\item{ego_id_name}{The name of the column with unique identifiers for
each ego.}
}
\value{
A vector of weights such that the observations with
these weights match the Washington Age / Race distribution and
regional MSM distribution.
}
\description{
This function allows for the reweighting of any dataset
with information on age (in 10 year increments from 15 to 65),
race (black, hispanic, or other),
and region of Washington to match the ACS based joint age / race
distribution and proportion of MSM in each region.
}
