% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{pd_demo_tidy}
\alias{pd_demo_tidy}
\alias{pd_ref_stock}
\alias{pd_ref_flows}
\alias{pd_ref_uasc_byage}
\alias{pd_ref_uasc_bysex}
\alias{pd_roc_stock}
\alias{pd_roc_flows}
\alias{pd_roc_uasc_byage}
\alias{pd_roc_uasc_bysex}
\alias{pd_rsd_meta}
\alias{pd_rsd_stock}
\alias{pd_rsd_flows}
\alias{pd_idp_stock}
\alias{pd_idp_flows}
\alias{pd_ret_flows}
\alias{pd_sta_stock}
\alias{pd_sta_flows}
\alias{pd_ooc_stock}
\alias{pd_ooc_flows}
\alias{pd_vda_stock}
\alias{pd_vda_flow}
\alias{pd_hst_stock}
\alias{pd_hst_flow}
\title{Convenience functions to tidy popdata tables}
\usage{
pd_demo_tidy(data)

pd_ref_stock(data)

pd_ref_flows(data)

pd_ref_uasc_byage(data)

pd_ref_uasc_bysex(data)

pd_roc_stock(data)

pd_roc_flows(data)

pd_roc_uasc_byage(data)

pd_roc_uasc_bysex(data)

pd_rsd_meta(data)

pd_rsd_stock(data)

pd_rsd_flows(data)

pd_idp_stock(data)

pd_idp_flows(data)

pd_ret_flows(data)

pd_sta_stock(data)

pd_sta_flows(data)

pd_ooc_stock(data)

pd_ooc_flows(data)

pd_vda_stock(data)

pd_vda_flow(data)

pd_hst_stock(data)

pd_hst_flow(data)
}
\arguments{
\item{data}{tibble, result of a call to \code{popdata()} or one of its wrappers.}
}
\value{
a tibble
}
\description{
Convenience functions to tidy popdata tables
}
