% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wa_acs_fullracebyregion.R
\docType{data}
\name{wa_acs_fullracebyregion}
\alias{wa_acs_fullracebyregion}
\title{Estimates of Washington race region joint distribution.}
\format{
A dataset with each row being a region and race, and the columns are
the region name, race, and number of males 15-65 in the specified race/region
group.  The corresponding percent is given in the last column.
}
\usage{
wa_acs_fullracebyregion
}
\description{
These estimates are based on ACS 2015 population estimates.
}
\details{
We are currently working to update documentation on the ACS data.  While
much of the data was taken from the ACS website American Fact Finder, this
system has been eliminated from the census's webpage, and it is still not clear
where to recover this data.
}
\keyword{datasets}
