% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wa_joint_age_race_dat.R
\docType{data}
\name{wa_joint_age_race_comp}
\alias{wa_joint_age_race_comp}
\alias{wa_joint_age_race}
\alias{wa_joint_age_race_comp,}
\title{ACS based estimates of age / race distributions in Washington.}
\usage{
wa_joint_age_race
}
\description{
ACS based estimates of age / race distributions in Washington.
}
\keyword{datasets}
