% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wa_joint_age_race_dat.R
\docType{data}
\name{wa_joint_age_race_comp}
\alias{wa_joint_age_race_comp}
\title{ACS based estimates of age / race distributions in Washington.}
\format{
An object of class \code{data.frame} with 15 rows and 4 columns.
}
\usage{
wa_joint_age_race_comp
}
\description{
These datasets are based on ACS information providing the joint
race age distribution of males in Washington.
The wa_joint_age_race_comp dataset is a compacted version of wa_joint_age_race
and only includes information from 15 - 65 in 10 year increments rather than
the 5 year increments between 0 and 85 that wa_joint_age_race has.
}
\details{
We are currently working to update documentation on the ACS data.  While
much of the data was taken from the ACS website American Fact Finder, this
system has been eliminated from the census's webpage, and it is still not clear
where to recover this data.
}
\keyword{datasets}
