% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_augment.R
\name{pd_augment}
\alias{pd_augment}
\title{Augment data-frame with country to region mapping using he country code package}
\usage{
pd_augment(data, expand)
}
\arguments{
\item{data}{tibble, dataset to augment -
previously obtained from \code{pd_asr} or \code{pd_mysr}}

\item{expand}{indicating either "asylum" or "origin"}
}
\value{
a tibble
}
\description{
Augment data-frame with country metadata using country code package -
https://github.com/vincentarelbundock/countrycode/blob/848e9ce363026c98a38daac9777005925f18f582/dictionary/get_unhcr.R
}
