% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdata-package.R, R/popdata.R
\docType{package}
\name{popdata}
\alias{popdata}
\alias{popdata-package}
\alias{pd_asr}
\alias{pd_mysr}
\alias{pd_pf}
\title{popdata}
\usage{
popdata(
  report = c("asr", "mysr", "pf"),
  table = c("refugees", "refugeeLike", "returnees", "demographics", "idp", "rsd",
    "stateless", "vda", "other", "hostcommunity", "coo", "ppg", "specific"),
  year = 2020
)

pd_asr(
  table = c("refugees", "refugeeLike", "returnees", "demographics", "idp", "rsd",
    "stateless", "vda", "other", "hostcommunity"),
  year = 2020
)

pd_mysr(
  table = c("refugees", "returnees", "idp", "rsd", "stateless", "other"),
  year = 2020
)

pd_pf(table = c("coo", "ppg", "specific"), year = 2020)
}
\arguments{
\item{report}{character, the type of report asr, mysr or pf}

\item{table}{character, the indicator refugees, refugeeLike, returnees,
demographics, idp, rsd, stateless, vda, other, hostcommunity, coo, ppg or specifics}

\item{year}{integer, the year}
}
\value{
A tibble
}
\description{
R client for UNHCR popdata

Popdata figures
}
\author{
\email{mail@ahmadoudicko.com}
}
\keyword{package}
