% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdata-package.R, R/popdata.R
\docType{package}
\name{popdata}
\alias{popdata}
\alias{popdata-package}
\alias{pd_asr}
\alias{pd_mysr}
\title{popdata}
\usage{
popdata(
  report = c("asr", "mysr"),
  table = c("refugees", "refugeeLike", "demographics", "rsd", "as2", "idp", "returnees",
    "stateless", "other", "oip", "hostCommunity", "comments"),
  year = 2022,
  quiet = getOption("popdata_quiet")
)

pd_asr(
  table = c("refugees", "refugeeLike", "demographics", "rsd", "as2", "idp", "returnees",
    "stateless", "other", "oip", "hostCommunity", "comments"),
  year = 2022,
  quiet = getOption("popdata_quiet")
)

pd_mysr(
  table = c("refugees", "refugeeLike", "demographics", "rsd", "as2", "idp", "returnees",
    "stateless", "other", "oip", "hostCommunity", "comments"),
  year = 2023,
  quiet = getOption("popdata_quiet")
)
}
\arguments{
\item{report}{character, the type of report asr, mysr}

\item{table}{character, any of the follwoing tables from within popdata:

\if{html}{\out{<div class="sourceCode">}}\preformatted{             *  "refugees", 
             *  "refugeeLike",
             *  "demographics",
             *  "rsd",
             *  "as2", 
             *  "idp",
             *  "returnees", 
             *  "stateless", 
             *  "other",
             *  "oip", 
             *  "hostCommunity",
             *  "comments"
}\if{html}{\out{</div>}}}

\item{year}{year for release}

\item{quiet}{option}
}
\value{
A tibble
}
\description{
R client for UNHCR popdata

Access UNHCR Popdata platform figures
}
\author{
\email{mail@ahmadoudicko.com}
}
\keyword{package}
